# docs/swagger.yaml
basePath: /api/v1/
definitions:
  controller._RespComment:
    properties:
      code:
        type: integer
      data:
        description: omitempty 当 data 为空时,不展示这个字段
      message:
        type: string
    type: object
  controller._ResponsePostList:
    properties:
      code:
        description: 业务响应状态码
        type: integer
      data:
        description: 数据
        items:
          $ref: '#/definitions/model.ApiPostDetail'
        type: array
      message:
        description: 提示信息
        type: string
    type: object
  model.ApiPostDetail:
    properties:
      author_id:
        type: integer
      author_name:
        type: string
      community_id:
        type: integer
      community_name:
        type: string
      content:
        type: string
      create_time:
        type: string
      introduction:
        description: omitempty 当 Introduction 为空时不展示
        type: string
      post_id:
        example: "0"
        type: string
      status:
        type: integer
      title:
        type: string
      vote_num:
        type: integer
    type: object
host: 192.168.24.133:8081
info:
  contact:
    email: support@swagger.io
    name: author:zorro
    url: http://www.swagger.io/support
  description: gin 框架开发 web 应用测试程序
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: bluebell
  version: "1.0"
paths:
  /comment:
    post:
      consumes:
      - application/json
      description: 创建评论接口
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: author_id
        type: integer
      - in: query
        name: comment_id
        type: integer
      - in: query
        name: content
        type: string
      - in: query
        name: create_time
        type: string
      - in: query
        name: parent_id
        type: integer
      - in: query
        name: question_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._RespComment'
      security:
      - ApiKeyAuth: []
      summary: 创建评论
      tags:
      - 评论
  /community:
    get:
      consumes:
      - application/json
      description: 社区列表
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: community_id
        type: integer
      - in: query
        name: community_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      security:
      - ApiKeyAuth: []
      summary: 社区列表
      tags:
      - 社区业务接口
  /community/:id:
    get:
      consumes:
      - application/json
      description: 社区详情
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: communityId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      security:
      - ApiKeyAuth: []
      summary: 社区详情
      tags:
      - 社区业务接口
  /posts2:
    get:
      consumes:
      - application/json
      description: 可按社区按时间或分数排序查询帖子列表接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 可以为空
        in: query
        name: community_id
        type: integer
      - description: 排序依据
        example: score
        in: query
        name: order
        type: string
      - description: 页码
        example: 1
        in: query
        name: page
        type: integer
      - description: 每页数量
        example: 10
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      security:
      - ApiKeyAuth: []
      summary: 升级版帖子列表接口
      tags:
      - 帖子相关接口
swagger: "2.0"